chore(data): move Daos of Clipboard and Draft and DbHelper to data
