+++ app/src/main/java/com/osfans/trime/data/db/DbHelper.java
@@ -1,4 +1,4 @@
-package com.osfans.trime.ime.symbol;
+package com.osfans.trime.data.db;
 
 import android.content.Context;
 import android.database.sqlite.SQLiteDatabase;
