+++ app/src/main/java/com/osfans/trime/ime/core/Trime.java
@@ -60,14 +60,14 @@ import androidx.annotation.Nullable;
 import androidx.appcompat.app.AlertDialog;
 import com.blankj.utilcode.util.BarUtils;
 import com.osfans.trime.R;
-import com.osfans.trime.clipboard.ClipboardDao;
+import com.osfans.trime.data.db.clipboard.ClipboardDao;
 import com.osfans.trime.common.ViewUtils;
 import com.osfans.trime.core.Rime;
 import com.osfans.trime.data.Config;
 import com.osfans.trime.data.AppPrefs;
 import com.osfans.trime.databinding.CompositionRootBinding;
 import com.osfans.trime.databinding.InputRootBinding;
-import com.osfans.trime.draft.DraftDao;
+import com.osfans.trime.data.db.draft.DraftDao;
 import com.osfans.trime.ime.enums.WindowsPositionType;
 import com.osfans.trime.ime.keyboard.Event;
 import com.osfans.trime.ime.keyboard.InputFeedbackManager;
